unified controller and hub versions to ensure successful local debugging
