+++ api/v1alpha1/edgex_conversion.go
@@ -13,4 +13,57 @@ limitations under the License.
 
 package v1alpha1
 
-func (*EdgeX) Hub() {}
+import (
+	"github.com/openyurtio/yurt-edgex-manager/api/v1alpha2"
+	corev1 "k8s.io/api/core/v1"
+	"sigs.k8s.io/controller-runtime/pkg/conversion"
+)
+
+func (src *EdgeX) ConvertTo(dstRaw conversion.Hub) error {
+	// Transform metadata
+	dst := dstRaw.(*v1alpha2.EdgeX)
+	dst.ObjectMeta = src.ObjectMeta
+	dst.TypeMeta = src.TypeMeta
+	dst.TypeMeta.APIVersion = "device.openyurt.io/v1alpha2"
+
+	// Transform spec
+	dst.Spec.Version = src.Spec.Version
+	dst.Spec.Security = false
+	dst.Spec.ImageRegistry = src.Spec.ImageRegistry
+	dst.Spec.PoolName = src.Spec.PoolName
+
+	// Transform status
+	dst.Status.Ready = src.Status.Ready
+	dst.Status.Initialized = src.Status.Initialized
+	dst.Status.ReadyComponentNum = src.Status.DeploymentReadyReplicas
+	dst.Status.UnreadyComponentNum = src.Status.DeploymentReplicas - src.Status.DeploymentReadyReplicas
+	dst.Status.Conditions = src.Status.Conditions
+
+	//TODO: Components
+	return nil
+}
+func (dst *EdgeX) ConvertFrom(srcRaw conversion.Hub) error {
+	// Transform metadata
+	src := srcRaw.(*v1alpha2.EdgeX)
+	dst.ObjectMeta = src.ObjectMeta
+	dst.TypeMeta = src.TypeMeta
+	dst.TypeMeta.APIVersion = "device.openyurt.io/v1alpha1"
+
+	// Transform spec
+	dst.Spec.Version = src.Spec.Version
+	dst.Spec.ImageRegistry = src.Spec.ImageRegistry
+	dst.Spec.PoolName = src.Spec.PoolName
+	dst.Spec.ServiceType = corev1.ServiceTypeClusterIP
+
+	// Transform status
+	dst.Status.Ready = src.Status.Ready
+	dst.Status.Initialized = src.Status.Initialized
+	dst.Status.ServiceReadyReplicas = src.Status.ReadyComponentNum
+	dst.Status.ServiceReplicas = src.Status.ReadyComponentNum + src.Status.UnreadyComponentNum
+	dst.Status.DeploymentReadyReplicas = src.Status.ReadyComponentNum
+	dst.Status.DeploymentReplicas = src.Status.ReadyComponentNum + src.Status.UnreadyComponentNum
+	dst.Status.Conditions = src.Status.Conditions
+
+	//TODO: AdditionalService and AdditionalDeployment
+	return nil
+}
